requirements.txt
pipecat-ai[anthropic,silero,daily,whisper,edge]
anthropic
python-dotenv
